Use relative / URL instead of site.url
Currently the website running on jsxgraph.org has invalid links on both `JSXGraph` link in the navbar and the logo. This is properly due to misconfiguration when building the website. In this PR I don't only report it, but also suggest an option to avoid the necessity to use site URL from the configuration.